lift(vc71): campaign match updates
Coverage: 35.5% (3295/9292 symbols)

Update permuter-campaign: path_add_directory & path_add_extension 98.0%→100.0%

Coverage: 35.5% (3295/9292 symbols)

Update permuter-campaign: accelerate_to_velocity3d 98.0%→100.0%

Coverage: 35.5% (3295/9292 symbols)

+++ src/halo/tag_files/files.c
@@ -389,7 +389,7 @@ void path_add_directory(char *path, const char *directory)
     return;
   }
 
-  if ((int)(csstrlen(path) + 1 + csstrlen(directory)) > 0xFF) {
+  if ((unsigned int)(csstrlen(path) + 1 + csstrlen(directory)) > 0xFF) {
     display_assert("strlen(path)+1+strlen(name)<=MAXIMUM_FILENAME_LENGTH",
                    "c:\\halo\\SOURCE\\tag_files\\files_windows.c", 0x2A0, true);
     system_exit(-1);
@@ -417,7 +417,7 @@ void path_add_extension(char *path, const char *extension)
     return;
   }
 
-  if ((int)csstrlen(path) + 1 + (int)csstrlen(extension) > 0xFF) {
+  if ((unsigned int)(csstrlen(path) + 1 + csstrlen(extension)) > 0xFF) {
     display_assert("strlen(path)+1+strlen(extension)<=MAXIMUM_FILENAME_LENGTH",
                    "c:\\halo\\SOURCE\\tag_files\\files_windows.c", 0x2b8, true);
     system_exit(-1);
